feat: extract common ui and logic
